Migrate PersonalDetails to TypeScript.
