+++ src/components/internal/PersonalDetails/ReadOnlyPersonalDetails.tsx
@@ -1,6 +1,6 @@
 import { h, Fragment } from 'preact';
 import Fieldset from '../FormFields/Fieldset';
-import { ReadOnlyPersonalDetailsProps } from './PersonalDetailsTypes';
+import { ReadOnlyPersonalDetailsProps } from './types';
 
 const ReadOnlyPersonalDetails = ({ data }) => {
     const { firstName, lastName, shopperEmail, telephoneNumber }: ReadOnlyPersonalDetailsProps = data;
