+++ src/components/internal/PersonalDetails/types.ts
@@ -1,0 +1,37 @@
+import { FieldsetVisibility, PersonalDetailsSchema } from '../../../types';
+
+export interface PersonalDetailsProps {
+    label?: string;
+    namePrefix?: string;
+    requiredFields?: string[];
+    visibility?: FieldsetVisibility;
+    data: PersonalDetailsSchema;
+    onChange: (newState: object) => void;
+    readonly?: boolean;
+    ref?: any;
+}
+
+export interface PersonalDetailsStateError {
+    firstName?: boolean;
+    lastName?: boolean;
+    gender?: boolean;
+    dateOfBirth?: boolean;
+    telephoneNumber?: boolean;
+    shopperEmail?: boolean;
+}
+
+export interface PersonalDetailsStateValid {
+    firstName?: boolean;
+    lastName?: boolean;
+    gender?: boolean;
+    dateOfBirth?: boolean;
+    telephoneNumber?: boolean;
+    shopperEmail?: boolean;
+}
+
+export interface ReadOnlyPersonalDetailsProps {
+    firstName?: string;
+    lastName?: string;
+    shopperEmail?: string;
+    telephoneNumber?: string;
+}
